added pagination and split components
