+++ packages/ui/components/pages/Fuse/FusePoolPage/MarketsList/index.tsx
@@ -1,62 +1,57 @@
 import {
-  ChevronDownIcon,
-  ChevronUpIcon,
-  ExternalLinkIcon,
-  LinkIcon,
-  QuestionIcon,
+  ChevronLeftIcon,
+  ChevronRightIcon,
+  TriangleDownIcon,
+  TriangleUpIcon,
 } from '@chakra-ui/icons';
 import {
   Box,
-  Button,
-  Link as ChakraLink,
+  Center,
+  Flex,
   HStack,
   IconButton,
-  Switch,
+  Select,
+  Table,
+  Tbody,
+  Td,
   Text,
-  VStack,
+  Th,
+  Thead,
+  Tr,
 } from '@chakra-ui/react';
-import { Web3Provider } from '@ethersproject/providers';
 import { FlywheelMarketRewardsInfo } from '@midas-capital/sdk/dist/cjs/src/modules/Flywheel';
-import { assetSymbols } from '@midas-capital/types';
-import { ColumnDef, Row } from '@tanstack/react-table';
-import { Contract, ContractTransaction, utils } from 'ethers';
-import LogRocket from 'logrocket';
+import {
+  ColumnDef,
+  flexRender,
+  getCoreRowModel,
+  getExpandedRowModel,
+  getPaginationRowModel,
+  getSortedRowModel,
+  PaginationState,
+  SortingState,
+  useReactTable,
+} from '@tanstack/react-table';
 import * as React from 'react';
-import { useEffect, useMemo, useState } from 'react';
+import { Fragment, useMemo } from 'react';
 
-import { DataTable } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/DataTable';
-import { RewardsInfo } from '@ui/components/pages/Fuse/FusePoolPage/SupplyList/RewardsInfo';
-import { CTokenIcon, TokenWithLabel } from '@ui/components/shared/CTokenIcon';
-import { Row as CRow } from '@ui/components/shared/Flex';
-import { PopoverTooltip } from '@ui/components/shared/PopoverTooltip';
-import { SimpleTooltip } from '@ui/components/shared/SimpleTooltip';
-import { SwitchCSS } from '@ui/components/shared/SwitchCSS';
-import {
-  aBNBcContractABI,
-  aBNBcContractAddress,
-  aprDays,
-  DOWN_LIMIT,
-  UP_LIMIT,
-  URL_MIDAS_DOCS,
-} from '@ui/constants/index';
-import { useMidas } from '@ui/context/MidasContext';
+import { AdditionalInfo } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/AdditionalInfo';
+import { BorrowApy } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/BorrowApy';
+import { BorrowBalance } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/BorrowBalance';
+import { Collateral } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/Collateral';
+import { ExpanderButton } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/ExpanderButton';
+import { Liquidity } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/Liquidity';
+import { Ltv } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/Ltv';
+import { Market } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/Market';
+import { SupplyApy } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/SupplyApy';
+import { SupplyBalance } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/SupplyBalance';
+import { Tvl } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/Tvl';
+import { FilterButton } from '@ui/components/shared/Button';
+import { MARKETS_COUNT_PER_PAGE } from '@ui/constants/index';
 import { useColors } from '@ui/hooks/useColors';
-import { usePluginInfo } from '@ui/hooks/usePluginInfo';
-import { useIsMobile } from '@ui/hooks/useScreenSize';
-import { useErrorToast, useInfoToast } from '@ui/hooks/useToast';
-import { useTokenData } from '@ui/hooks/useTokenData';
 import { MarketData } from '@ui/types/TokensDataMap';
-import {
-  aprFormatter,
-  shortUsdFormatter,
-  smallUsdFormatter,
-  tokenFormatter,
-} from '@ui/utils/bigUtils';
-import { errorCodeToMessage } from '@ui/utils/errorCodeToMessage';
-import { getBlockTimePerMinuteByChainId } from '@ui/utils/networkData';
 import { sortAssets } from '@ui/utils/sortAssets';
 
-type Market = {
+export type Market = {
   market: MarketData;
   supplyApy: MarketData;
   supplyBalance: MarketData;
@@ -66,455 +61,6 @@ type Market = {
   ltv: MarketData;
   borrowBalance: MarketData;
   liquidity: MarketData;
-};
-
-const ExpanderComponent = ({
-  getToggleExpandedHandler,
-  isExpanded,
-  canExpand,
-}: {
-  getToggleExpandedHandler: () => void;
-  isExpanded: boolean;
-  canExpand: boolean;
-}) => {
-  const { cOutlineBtn } = useColors();
-
-  return (
-    <IconButton
-      alignSelf="flex-end"
-      onClick={(e) => {
-        e.stopPropagation();
-        getToggleExpandedHandler();
-      }}
-      disabled={!canExpand ? true : false}
-      variant="outline"
-      color={isExpanded ? cOutlineBtn.primary.hoverTxtColor : cOutlineBtn.primary.txtColor}
-      aria-label="detail view"
-      borderRadius="50%"
-      borderWidth={3}
-      borderColor={cOutlineBtn.primary.borderColor}
-      background={isExpanded ? cOutlineBtn.primary.hoverBgColor : cOutlineBtn.primary.bgColor}
-      icon={isExpanded ? <ChevronUpIcon fontSize={30} /> : <ChevronDownIcon fontSize={30} />}
-      _hover={{
-        background: cOutlineBtn.primary.hoverBgColor,
-        color: cOutlineBtn.primary.hoverTxtColor,
-      }}
-    />
-  );
-};
-
-const LiquidityComponent = ({ asset }: { asset: MarketData }) => {
-  const { data: tokenData } = useTokenData(asset.underlyingToken);
-  const { cCard } = useColors();
-
-  return (
-    <Box textAlign="end">
-      <PopoverTooltip
-        body={
-          <>
-            {asset.liquidityFiat > DOWN_LIMIT && asset.liquidityFiat < UP_LIMIT && (
-              <>
-                <div>{asset.liquidityFiat.toString()}</div>
-                <br />
-              </>
-            )}
-            <div>
-              Liquidity is the amount of this asset that is available to borrow (unborrowed). To see
-              how much has been supplied and borrowed in total, navigate to the Pool Info tab.
-            </div>
-          </>
-        }
-        placement="top-end"
-      >
-        <VStack alignItems={'flex-end'}>
-          <Text color={cCard.txtColor} fontWeight="bold" fontSize={{ base: '2.8vw', sm: 'md' }}>
-            {smallUsdFormatter(asset.liquidityFiat)}
-            {asset.liquidityFiat > DOWN_LIMIT && asset.liquidityFiat < UP_LIMIT && '+'}
-          </Text>
-          <Text color={cCard.txtColor} fontSize={{ base: '2.8vw', sm: '0.8rem' }}>
-            {shortUsdFormatter(
-              Number(utils.formatUnits(asset.liquidity, asset.underlyingDecimals))
-            ).replace('$', '')}{' '}
-            {tokenData?.symbol}
-          </Text>
-        </VStack>
-      </PopoverTooltip>
-    </Box>
-  );
-};
-
-const BorrowBalanceComponent = ({ asset }: { asset: MarketData }) => {
-  const { data: tokenData } = useTokenData(asset.underlyingToken);
-  const { cCard } = useColors();
-
-  return (
-    <VStack alignItems={'flex-end'}>
-      <SimpleTooltip
-        label={asset.borrowBalanceFiat.toString()}
-        isDisabled={asset.borrowBalanceFiat === DOWN_LIMIT || asset.borrowBalanceFiat >= UP_LIMIT}
-      >
-        <Text color={cCard.txtColor} fontWeight="bold" fontSize={{ base: '2.8vw', sm: 'md' }}>
-          {smallUsdFormatter(asset.borrowBalanceFiat)}
-          {asset.borrowBalanceFiat > DOWN_LIMIT && asset.borrowBalanceFiat < UP_LIMIT && '+'}
-        </Text>
-      </SimpleTooltip>
-      <SimpleTooltip
-        label={utils.formatUnits(asset.borrowBalance, asset.underlyingDecimals)}
-        isDisabled={
-          Number(utils.formatUnits(asset.borrowBalance, asset.underlyingDecimals)) === DOWN_LIMIT ||
-          Number(utils.formatUnits(asset.borrowBalance, asset.underlyingDecimals)) >= UP_LIMIT
-        }
-      >
-        <Text color={cCard.txtColor} mt={1} fontSize={{ base: '2.8vw', sm: '0.8rem' }}>
-          {smallUsdFormatter(
-            Number(utils.formatUnits(asset.borrowBalance, asset.underlyingDecimals))
-          ).replace('$', '')}
-          {Number(utils.formatUnits(asset.borrowBalance, asset.underlyingDecimals)) > DOWN_LIMIT &&
-            Number(utils.formatUnits(asset.borrowBalance, asset.underlyingDecimals)) < UP_LIMIT &&
-            '+'}{' '}
-          {tokenData?.symbol ?? asset.underlyingSymbol}
-        </Text>
-      </SimpleTooltip>
-    </VStack>
-  );
-};
-
-const LtvComponent = ({ asset }: { asset: MarketData }) => {
-  const { cCard } = useColors();
-
-  return (
-    <Box textAlign="end">
-      <PopoverTooltip
-        placement="top-start"
-        body={
-          'The Loan to Value (LTV) ratio defines the maximum amount of tokens in the pool that can be borrowed with a specific collateral. It’s expressed in percentage: if in a pool ETH has 75% LTV, for every 1 ETH worth of collateral, borrowers will be able to borrow 0.75 ETH worth of other tokens in the pool.'
-        }
-      >
-        <Text color={cCard.txtColor} fontSize={{ base: '2.8vw', sm: 'lg' }}>
-          {utils.formatUnits(asset.collateralFactor, 16)}%
-        </Text>
-      </PopoverTooltip>
-    </Box>
-  );
-};
-
-const TvlComponent = ({ asset }: { asset: MarketData }) => {
-  const { cCard } = useColors();
-
-  return (
-    <VStack alignItems={'flex-end'}>
-      <PopoverTooltip
-        placement="top-start"
-        body={
-          "Total Value Lent (TVL) measures how much of this asset has been supplied in total. TVL does not account for how much of the lent assets have been borrowed, use 'liquidity' to determine the total unborrowed assets lent."
-        }
-      >
-        <Text wordBreak={'keep-all'} color={cCard.txtColor} fontSize={{ base: '2.8vw', sm: 'lg' }}>
-          {shortUsdFormatter(asset.totalSupplyFiat)}
-        </Text>
-      </PopoverTooltip>
-    </VStack>
-  );
-};
-
-const BorrowApyComponent = ({ asset }: { asset: MarketData }) => {
-  const { currentChain, midasSdk } = useMidas();
-  const blocksPerMin = getBlockTimePerMinuteByChainId(currentChain.id);
-  const borrowAPR = midasSdk.ratePerBlockToAPY(asset.borrowRatePerBlock, blocksPerMin);
-  const { cCard } = useColors();
-
-  return (
-    <VStack alignItems={'flex-end'}>
-      <Text color={cCard.txtColor} fontSize={{ base: '2.8vw', sm: 'lg' }}>
-        {borrowAPR.toFixed(3)}%
-      </Text>
-    </VStack>
-  );
-};
-
-const CollateralComponent = ({
-  asset,
-  comptrollerAddress,
-}: {
-  asset: MarketData;
-  comptrollerAddress: string;
-}) => {
-  const { midasSdk, setPendingTxHash } = useMidas();
-  const errorToast = useErrorToast();
-  const infoToast = useInfoToast();
-
-  const { cSwitch } = useColors();
-  const isMobile = useIsMobile();
-
-  const onToggleCollateral = async () => {
-    const comptroller = midasSdk.createComptroller(comptrollerAddress);
-
-    let call: ContractTransaction;
-    if (asset.membership) {
-      const exitCode = await comptroller.callStatic.exitMarket(asset.cToken);
-      if (!exitCode.eq(0)) {
-        infoToast({
-          title: 'Cannot Remove Collateral',
-          description: errorCodeToMessage(exitCode.toNumber()),
-        });
-        return;
-      }
-      call = await comptroller.exitMarket(asset.cToken);
-    } else {
-      call = await comptroller.enterMarkets([asset.cToken]);
-    }
-
-    if (!call) {
-      if (asset.membership) {
-        errorToast({
-          title: 'Error! Code: ' + call,
-          description:
-            'You cannot disable this asset as collateral as you would not have enough collateral posted to keep your borrow. Try adding more collateral of another type or paying back some of your debt.',
-        });
-      } else {
-        errorToast({
-          title: 'Error! Code: ' + call,
-          description: 'You cannot enable this asset as collateral at this time.',
-        });
-      }
-
-      return;
-    }
-
-    setPendingTxHash(call.hash);
-
-    LogRocket.track('Fuse-ToggleCollateral');
-  };
-  return (
-    <CRow mainAxisAlignment="flex-end" crossAxisAlignment="center">
-      <SwitchCSS symbol={asset.underlyingSymbol.replace(/[\s+()]/g, '')} color={cSwitch.bgColor} />
-      <Switch
-        isChecked={asset.membership}
-        className={'switch-' + asset.underlyingSymbol.replace(/[\s+()]/g, '')}
-        onChange={onToggleCollateral}
-        size={isMobile ? 'sm' : 'md'}
-        cursor={'pointer'}
-      />
-    </CRow>
-  );
-};
-
-const SupplyBalanceComponent = ({ asset }: { asset: MarketData }) => {
-  const { data: tokenData } = useTokenData(asset.underlyingToken);
-
-  const { cCard } = useColors();
-
-  return (
-    <VStack alignItems="flex-end">
-      <SimpleTooltip
-        label={asset.supplyBalanceFiat.toString()}
-        isDisabled={asset.supplyBalanceFiat === DOWN_LIMIT || asset.supplyBalanceFiat >= UP_LIMIT}
-      >
-        <Text color={cCard.txtColor} fontWeight="bold" fontSize={{ base: '2.8vw', sm: 'md' }}>
-          {smallUsdFormatter(asset.supplyBalanceFiat)}
-          {asset.supplyBalanceFiat > DOWN_LIMIT && asset.supplyBalanceFiat < UP_LIMIT && '+'}
-        </Text>
-      </SimpleTooltip>
-      <SimpleTooltip
-        label={utils.formatUnits(asset.supplyBalance, asset.underlyingDecimals)}
-        isDisabled={
-          Number(utils.formatUnits(asset.supplyBalance, asset.underlyingDecimals)) === DOWN_LIMIT ||
-          Number(utils.formatUnits(asset.supplyBalance, asset.underlyingDecimals)) >= UP_LIMIT
-        }
-      >
-        <Text color={cCard.txtColor} mt={1} fontSize={{ base: '2.8vw', sm: '0.8rem' }}>
-          {tokenFormatter(asset.supplyBalance, asset.underlyingDecimals)}
-          {Number(utils.formatUnits(asset.supplyBalance, asset.underlyingDecimals)) > DOWN_LIMIT &&
-            Number(utils.formatUnits(asset.supplyBalance, asset.underlyingDecimals)) < UP_LIMIT &&
-            '+'}{' '}
-          {tokenData?.extraData?.shortName ?? tokenData?.symbol ?? asset.underlyingSymbol}
-        </Text>
-      </SimpleTooltip>
-    </VStack>
-  );
-};
-
-const SupplyApyComponent = ({
-  asset,
-  rewards,
-}: {
-  asset: MarketData;
-  rewards: FlywheelMarketRewardsInfo[];
-}) => {
-  const { midasSdk, currentChain } = useMidas();
-  const supplyAPY = midasSdk.ratePerBlockToAPY(
-    asset.supplyRatePerBlock,
-    getBlockTimePerMinuteByChainId(currentChain.id)
-  );
-
-  const { cCard } = useColors();
-
-  const rewardsOfThisMarket = useMemo(
-    () => rewards.find((r) => r.market === asset.cToken),
-    [asset.cToken, rewards]
-  );
-
-  const [aBNBcApr, setaBNBcApr] = useState('');
-
-  useEffect(() => {
-    const func = async () => {
-      const contract = new Contract(
-        aBNBcContractAddress,
-        aBNBcContractABI,
-        midasSdk.provider as Web3Provider
-      );
-
-      const apr = await contract.callStatic.averagePercentageRate(aprDays);
-      setaBNBcApr(utils.formatUnits(apr));
-    };
-
-    if (asset.underlyingSymbol === assetSymbols.aBNBc) {
-      func();
-    }
-  }, [asset, midasSdk.provider]);
-
-  return (
-    <VStack alignItems={'flex-end'}>
-      <Text color={cCard.txtColor} fontWeight="bold" fontSize={{ base: '2.8vw', sm: 'md' }}>
-        {supplyAPY.toFixed(2)}%
-      </Text>
-      {asset.underlyingSymbol === assetSymbols.aBNBc && (
-        <Text color={cCard.txtColor} fontSize={{ base: '2.8vw', sm: 'md' }}>
-          + {Number(aBNBcApr).toFixed(2)}%
-        </Text>
-      )}
-
-      {rewardsOfThisMarket?.rewardsInfo && rewardsOfThisMarket?.rewardsInfo.length !== 0 ? (
-        rewardsOfThisMarket?.rewardsInfo.map((info) =>
-          asset.plugin ? (
-            <RewardsInfo
-              key={info.rewardToken}
-              underlyingAddress={asset.underlyingToken}
-              pluginAddress={asset.plugin}
-              rewardAddress={info.rewardToken}
-            />
-          ) : (
-            <HStack key={info.rewardToken} justifyContent={'flex-end'} spacing={0}>
-              <HStack mr={2}>
-                <Text fontSize={{ base: '3.2vw', sm: '0.9rem' }}>+</Text>
-                <TokenWithLabel address={info.rewardToken} size="2xs" />
-              </HStack>
-              {info.formattedAPR && (
-                <Text color={cCard.txtColor} fontSize={{ base: '2.8vw', sm: '0.8rem' }} ml={1}>
-                  {aprFormatter(info.formattedAPR)}%
-                </Text>
-              )}
-            </HStack>
-          )
-        )
-      ) : asset.plugin ? (
-        <RewardsInfo underlyingAddress={asset.underlyingToken} pluginAddress={asset.plugin} />
-      ) : null}
-    </VStack>
-  );
-};
-
-const MarketComponent = ({ asset }: { asset: MarketData }) => {
-  const { scanUrl } = useMidas();
-  const { data: tokenData } = useTokenData(asset.underlyingToken);
-
-  const { cCard } = useColors();
-
-  const { data: pluginInfo } = usePluginInfo(asset.plugin);
-
-  return (
-    <CRow mainAxisAlignment="flex-start" crossAxisAlignment="center">
-      <CTokenIcon size="sm" address={asset.underlyingToken} />
-      <VStack alignItems={'flex-start'} ml={2}>
-        <PopoverTooltip
-          placement="top-start"
-          body={
-            <div
-              dangerouslySetInnerHTML={{
-                __html: asset.extraDocs || asset.underlyingSymbol,
-              }}
-            />
-          }
-        >
-          <Text fontWeight="bold" textAlign={'left'} fontSize={{ base: '2.8vw', sm: '0.9rem' }}>
-            {tokenData?.symbol ?? asset.underlyingSymbol}
-          </Text>
-        </PopoverTooltip>
-      </VStack>
-
-      <HStack ml={2}>
-        {asset.underlyingSymbol &&
-          tokenData?.symbol &&
-          asset.underlyingSymbol.toLowerCase() !== tokenData?.symbol?.toLowerCase() && (
-            <PopoverTooltip body={asset.underlyingSymbol}>
-              <QuestionIcon />
-            </PopoverTooltip>
-          )}
-        <PopoverTooltip placement="top-start" body={`${scanUrl}/address/${asset.underlyingToken}`}>
-          <Button
-            minWidth={6}
-            m={0}
-            variant={'link'}
-            as={ChakraLink}
-            href={`${scanUrl}/address/${asset.underlyingToken}`}
-            isExternal
-            onClick={(e) => {
-              e.stopPropagation();
-            }}
-          >
-            <LinkIcon h={{ base: 3, sm: 6 }} color={cCard.txtColor} />
-          </Button>
-        </PopoverTooltip>
-
-        {asset.plugin && (
-          <PopoverTooltip
-            placement="top-start"
-            body={
-              <Text lineHeight="base">
-                This market is using the <b>{pluginInfo?.name}</b> ERC4626 Strategy.
-                <br />
-                {pluginInfo?.apyDocsUrl ? (
-                  <ChakraLink
-                    href={pluginInfo.apyDocsUrl}
-                    isExternal
-                    variant={'color'}
-                    onClick={(e) => {
-                      e.stopPropagation();
-                    }}
-                  >
-                    Vault details
-                  </ChakraLink>
-                ) : (
-                  <>
-                    Read more about it{' '}
-                    <ChakraLink
-                      href={pluginInfo?.strategyDocsUrl || URL_MIDAS_DOCS}
-                      isExternal
-                      variant={'color'}
-                      onClick={(e) => {
-                        e.stopPropagation();
-                      }}
-                    >
-                      in our Docs <ExternalLinkIcon mx="2px" />
-                    </ChakraLink>
-                  </>
-                )}
-                .
-              </Text>
-            }
-          >
-            <span role="img" aria-label="plugin" style={{ fontSize: 18 }}>
-              🔌
-            </span>
-          </PopoverTooltip>
-        )}
-      </HStack>
-    </CRow>
-  );
-};
-
-const renderSubComponent = ({ row }: { row: Row<Market> }) => {
-  return <pre style={{ fontSize: '20px' }}>{row && <code>Additional info here</code>}</pre>;
 };
 
 export const MarketsList = ({
@@ -554,145 +100,283 @@ export const MarketsList = ({
   const columns: ColumnDef<Market>[] = useMemo(() => {
     return [
       {
-        id: 'assets',
+        accessorKey: 'market',
         header: () => (
-          <Text textAlign="start" py={2} fontSize="18">
-            Assets
+          <Text textAlign="start" py={2}>
+            Market
+          </Text>
+        ),
+        cell: ({ getValue }) => <Market asset={getValue<MarketData>()} />,
+        footer: (props) => props.column.id,
+      },
+      {
+        accessorFn: (row) => row.ltv,
+        id: 'LTV',
+        cell: ({ getValue }) => <Ltv asset={getValue<MarketData>()} />,
+        header: () => (
+          <Text textAlign="end" py={2}>
+            LTV
           </Text>
         ),
         footer: (props) => props.column.id,
-        columns: [
-          {
-            accessorKey: 'market',
-            header: () => (
-              <Text textAlign="start" py={2}>
-                Market
-              </Text>
-            ),
-            cell: ({ getValue }) => <MarketComponent asset={getValue<MarketData>()} />,
-            footer: (props) => props.column.id,
-          },
-          {
-            accessorFn: (row) => row.ltv,
-            id: 'LTV',
-            cell: ({ getValue }) => <LtvComponent asset={getValue<MarketData>()} />,
-            header: () => (
-              <Text textAlign="end" py={2}>
-                LTV
-              </Text>
-            ),
-            footer: (props) => props.column.id,
-          },
-          {
-            accessorFn: (row) => row.supplyApy,
-            id: 'supplyApy',
-            cell: ({ getValue }) => (
-              <SupplyApyComponent asset={getValue<MarketData>()} rewards={rewards} />
-            ),
-            header: () => (
-              <Text textAlign="end" py={2}>
-                Supply APY
-              </Text>
-            ),
-            footer: (props) => props.column.id,
-          },
-          {
-            accessorFn: (row) => row.borrowApy,
-            id: 'borrowApy',
-            cell: ({ getValue }) => <BorrowApyComponent asset={getValue<MarketData>()} />,
-            header: () => (
-              <Text textAlign="end" py={2}>
-                Borrow Apy
-              </Text>
-            ),
-            footer: (props) => props.column.id,
-          },
-          {
-            accessorFn: (row) => row.tvl,
-            id: 'tvl',
-            cell: ({ getValue }) => <TvlComponent asset={getValue<MarketData>()} />,
-            header: () => (
-              <Text textAlign="end" py={2}>
-                TVL
-              </Text>
-            ),
-            footer: (props) => props.column.id,
-          },
-          {
-            accessorFn: (row) => row.supplyBalance,
-            id: 'supplyBalance',
-            cell: ({ getValue }) => <SupplyBalanceComponent asset={getValue<MarketData>()} />,
-            header: () => (
-              <Text textAlign="end" py={2}>
-                Supply Balance
-              </Text>
-            ),
-            footer: (props) => props.column.id,
-          },
-          {
-            accessorFn: (row) => row.borrowBalance,
-            id: 'borrowBalance',
-            cell: ({ getValue }) => <BorrowBalanceComponent asset={getValue<MarketData>()} />,
-            header: () => (
-              <Text textAlign="end" py={2}>
-                Borrow Balance
-              </Text>
-            ),
-            footer: (props) => props.column.id,
-          },
-          {
-            accessorFn: (row) => row.liquidity,
-            id: 'liquidity',
-            cell: ({ getValue }) => <LiquidityComponent asset={getValue<MarketData>()} />,
-            header: () => (
-              <Text textAlign="end" py={2}>
-                Liquidity
-              </Text>
-            ),
-            footer: (props) => props.column.id,
-          },
-          {
-            accessorFn: (row) => row.collateral,
-            id: 'collateral',
-            cell: ({ getValue }) => (
-              <CollateralComponent
-                asset={getValue<MarketData>()}
-                comptrollerAddress={comptrollerAddress}
-              />
-            ),
-            header: () => (
-              <Text textAlign="end" py={2}>
-                Collateral
-              </Text>
-            ),
-            footer: (props) => props.column.id,
-          },
-          {
-            id: 'expander',
-            header: () => null,
-            cell: ({ row }) => {
-              return (
-                <ExpanderComponent
-                  getToggleExpandedHandler={row.getToggleExpandedHandler()}
-                  isExpanded={row.getIsExpanded()}
-                  canExpand={row.getCanExpand()}
-                />
-              );
-            },
-          },
-        ],
+      },
+      {
+        accessorFn: (row) => row.supplyApy,
+        id: 'supplyApy',
+        cell: ({ getValue }) => <SupplyApy asset={getValue<MarketData>()} rewards={rewards} />,
+        header: () => (
+          <Text textAlign="end" py={2}>
+            Supply APY
+          </Text>
+        ),
+        footer: (props) => props.column.id,
+      },
+      {
+        accessorFn: (row) => row.borrowApy,
+        id: 'borrowApy',
+        cell: ({ getValue }) => <BorrowApy asset={getValue<MarketData>()} />,
+        header: () => (
+          <Text textAlign="end" py={2}>
+            Borrow Apy
+          </Text>
+        ),
+        footer: (props) => props.column.id,
+      },
+      {
+        accessorFn: (row) => row.tvl,
+        id: 'tvl',
+        cell: ({ getValue }) => <Tvl asset={getValue<MarketData>()} />,
+        header: () => (
+          <Text textAlign="end" py={2}>
+            TVL
+          </Text>
+        ),
+        footer: (props) => props.column.id,
+      },
+      {
+        accessorFn: (row) => row.supplyBalance,
+        id: 'supplyBalance',
+        cell: ({ getValue }) => <SupplyBalance asset={getValue<MarketData>()} />,
+        header: () => (
+          <Text textAlign="end" py={2}>
+            Supply Balance
+          </Text>
+        ),
+        footer: (props) => props.column.id,
+      },
+      {
+        accessorFn: (row) => row.borrowBalance,
+        id: 'borrowBalance',
+        cell: ({ getValue }) => <BorrowBalance asset={getValue<MarketData>()} />,
+        header: () => (
+          <Text textAlign="end" py={2}>
+            Borrow Balance
+          </Text>
+        ),
+        footer: (props) => props.column.id,
+      },
+      {
+        accessorFn: (row) => row.liquidity,
+        id: 'liquidity',
+        cell: ({ getValue }) => <Liquidity asset={getValue<MarketData>()} />,
+        header: () => (
+          <Text textAlign="end" py={2}>
+            Liquidity
+          </Text>
+        ),
+        footer: (props) => props.column.id,
+      },
+      {
+        accessorFn: (row) => row.collateral,
+        id: 'collateral',
+        cell: ({ getValue }) => (
+          <Collateral asset={getValue<MarketData>()} comptrollerAddress={comptrollerAddress} />
+        ),
+        header: () => (
+          <Text textAlign="end" py={2}>
+            Collateral
+          </Text>
+        ),
+        footer: (props) => props.column.id,
+      },
+      {
+        id: 'expander',
+        header: () => null,
+        cell: ({ row }) => {
+          return (
+            <ExpanderButton
+              getToggleExpandedHandler={row.getToggleExpandedHandler()}
+              isExpanded={row.getIsExpanded()}
+              canExpand={row.getCanExpand()}
+            />
+          );
+        },
       },
     ];
   }, [rewards, comptrollerAddress]);
 
+  const [sorting, setSorting] = React.useState<SortingState>([]);
+  const [pagination, onPagination] = React.useState<PaginationState>({
+    pageIndex: 0,
+    pageSize: 5,
+  });
+  const table = useReactTable({
+    columns,
+    data,
+    getRowCanExpand: () => true,
+    getPaginationRowModel: getPaginationRowModel(),
+    onPaginationChange: onPagination,
+    getCoreRowModel: getCoreRowModel(),
+    onSortingChange: setSorting,
+    getSortedRowModel: getSortedRowModel(),
+    getExpandedRowModel: getExpandedRowModel(),
+    state: {
+      sorting,
+      pagination,
+    },
+  });
+
+  const { cCard } = useColors();
+
   return (
     <Box overflowX="auto">
-      <DataTable
-        columns={columns}
-        data={data}
-        getRowCanExpand={() => true}
-        renderSubComponent={renderSubComponent}
-      />
+      <Flex justifyContent="space-between" px="4" py="8">
+        <HStack className="pagination" gap={4}>
+          <Text fontSize={24}>Assets</Text>
+          <FilterButton
+            variant="filter"
+            isSelected={true}
+            onClick={() => {
+              // setSelectedAsset(asset);
+              // setSelectedIndex(index);
+            }}
+            px={2}
+          >
+            <Center px={1} fontWeight="bold">
+              Collateral
+            </Center>
+          </FilterButton>
+        </HStack>
+        <HStack className="pagination" gap={4}>
+          <HStack>
+            <Text>Rows Per Page :</Text>
+            <Select
+              value={pagination.pageSize}
+              onChange={(e) => {
+                table.setPageSize(Number(e.target.value));
+              }}
+              maxW="max-content"
+            >
+              {MARKETS_COUNT_PER_PAGE.map((pageSize) => (
+                <option key={pageSize} value={pageSize}>
+                  {pageSize}
+                </option>
+              ))}
+            </Select>
+          </HStack>
+          <HStack gap={2}>
+            <Text>
+              {pagination.pageIndex * pagination.pageSize + 1} -{' '}
+              {(pagination.pageIndex + 1) * pagination.pageSize >
+              table.getCoreRowModel().rows.length
+                ? table.getCoreRowModel().rows.length
+                : (pagination.pageIndex + 1) * pagination.pageSize}{' '}
+              of {table.getCoreRowModel().rows.length}
+            </Text>
+            <HStack>
+              <IconButton
+                variant="outline"
+                aria-label="toPrevious"
+                icon={<ChevronLeftIcon fontSize={30} />}
+                onClick={() => table.previousPage()}
+                disabled={!table.getCanPreviousPage()}
+                isRound
+              />
+              <IconButton
+                variant="outline"
+                aria-label="toNext"
+                icon={<ChevronRightIcon fontSize={30} />}
+                onClick={() => table.nextPage()}
+                disabled={!table.getCanNextPage()}
+                isRound
+              />
+            </HStack>
+          </HStack>
+        </HStack>
+      </Flex>
+
+      <Table>
+        <Thead>
+          {table.getHeaderGroups().map((headerGroup) => (
+            <Tr
+              key={headerGroup.id}
+              borderColor={cCard.dividerColor}
+              borderBottomWidth={1}
+              borderTopWidth={2}
+            >
+              {headerGroup.headers.map((header) => {
+                return (
+                  <Th
+                    key={header.id}
+                    onClick={header.column.getToggleSortingHandler()}
+                    border="none"
+                    color={cCard.txtColor}
+                    fontSize={16}
+                    textTransform="capitalize"
+                    py={4}
+                  >
+                    <HStack gap={1} justifyContent={header.index === 0 ? 'flex-start' : 'flex-end'}>
+                      <>{flexRender(header.column.columnDef.header, header.getContext())}</>
+                      <>
+                        {header.column.getIsSorted() ? (
+                          header.column.getIsSorted() === 'desc' ? (
+                            <TriangleDownIcon aria-label="sorted descending" />
+                          ) : (
+                            <TriangleUpIcon aria-label="sorted ascending" />
+                          )
+                        ) : null}
+                      </>
+                    </HStack>
+                  </Th>
+                );
+              })}
+            </Tr>
+          ))}
+        </Thead>
+        <Tbody>
+          {table.getRowModel().rows.map((row) => (
+            <Fragment key={row.id}>
+              <Tr
+                key={row.id}
+                borderColor={cCard.dividerColor}
+                borderTopWidth={row.getIsExpanded() ? 4 : 1}
+                background={row.getIsExpanded() ? cCard.hoverBgColor : cCard.bgColor}
+                _hover={{ bg: cCard.hoverBgColor }}
+              >
+                {row.getVisibleCells().map((cell) => {
+                  return (
+                    <Td key={cell.id} border="none">
+                      {flexRender(cell.column.columnDef.cell, cell.getContext())}
+                    </Td>
+                  );
+                })}
+              </Tr>
+              {row.getIsExpanded() && (
+                <Tr
+                  borderColor={cCard.dividerColor}
+                  borderBottomWidth={row.getIsExpanded() ? 6 : 0}
+                  background={row.getIsExpanded() ? cCard.hoverBgColor : cCard.bgColor}
+                >
+                  {/* 2nd row is a custom 1 cell row */}
+                  <Td border="none" colSpan={row.getVisibleCells().length}>
+                    <AdditionalInfo row={row} />
+                  </Td>
+                </Tr>
+              )}
+            </Fragment>
+          ))}
+        </Tbody>
+      </Table>
     </Box>
   );
 };
