+++ packages/ui/constants/index.ts
@@ -74,3 +74,4 @@ export const aBNBcContractABI = [
 ];
 
 export const DISCORD_URL = 'https://discord.gg/85YxVuPeMt';
+export const MARKETS_COUNT_PER_PAGE = [5, 10, 20, 30, 50];
